Update README.md
Autonomous finance is a revolutionary approach in the rapidly evolving world of technological and augmented tools that CFOs and heads of financial planning and analysis (FP&A) have at their disposal. Coined by industry analyst Gartner, it refers to self-learning software agents that automate business operations and corporate finances. Autonomous finance is increasingly permeating all aspects of financial management by optimizing and streamlining financial data gathering and analysis.
READ MORE-https://www.iconresources.com/embracing-the-future-the-rise-of-autonomous-finance-in-organizations/